requirements.txt
pandas>=1.5.0
numpy>=1.24.0
scikit-learn>=1.2.0
matplotlib>=3.7.0
pyyaml>=6.0
mlxtend>=0.21.0
papermill>=2.5.0
jupyter>=1.0.0
joblib>=1.2.0
